added delete folder after archiving

+++ src/DeleteDir.java
@@ -4,8 +4,11 @@ public class DeleteDir {
     public static void deleteDirectory(File dir) {
         if (dir.isDirectory()) {
             String[] children = dir.list();
-            for (int i = 0; i < children.length; i++) {
-                File f = new File(dir, children[i]);
+            if(children==null){
+                return;
+            }
+            for (String child : children) {
+                File f = new File(dir, child);
                 deleteDirectory(f);
             }
         }
